Final OXO-Game
I developed this game in my first year.

The "OXO game," also known as "Tic-Tac-Toe," is a classic two-player game played on a 3x3 grid. The grid is initially empty, and the players take turns marking an empty cell with their respective symbols (usually "X" for one player and "O" for the other). The goal is to create a line of three of your symbols horizontally, vertically, or diagonally. The game ends when one player achieves this goal, or if the entire grid is filled without a winner, resulting in a draw. Tic-Tac-Toe is simple yet strategic, making it a popular and quick game to play